Pull request 151: osutil-executil
Merge in GO/golibs from osutil-executil to master

Squashed commit of the following:

commit 4857d8ff2f4b47aac65b529e78c370281bce92a3
Author: Ainar Garipov <A.Garipov@AdGuard.COM>
Date:   Thu Jul 31 16:52:43 2025 +0300

    osutil/executil: rename methods

commit 0c65ec13b63dc13cb45f5c1d81d59cdbe7caf9ed
Author: Ainar Garipov <A.Garipov@AdGuard.COM>
Date:   Thu Jul 31 16:13:25 2025 +0300

    all: imp docs, tests

commit e3fc16d991e0bc2c79f98befd7de819061812e08
Author: Ainar Garipov <A.Garipov@AdGuard.COM>
Date:   Wed Jul 30 20:42:49 2025 +0300

    bamboo-specs: upd

commit 928465595acf725379620d240d5f1aec7ebba0e2
Author: Ainar Garipov <A.Garipov@AdGuard.COM>
Date:   Wed Jul 30 20:35:34 2025 +0300

    all: rm exe; imp gitignore; fmt

commit 504966783fd26f7cf10f7eddd95e9c8c18878eca
Author: Ainar Garipov <A.Garipov@AdGuard.COM>
Date:   Wed Jul 30 20:24:12 2025 +0300

    osutil: add executil

+++ ioutil/truncwriter.go
@@ -6,13 +6,13 @@ import "io"
 // to its underlying writer and then ignores the rest.
 type TruncatedWriter struct {
 	w      io.Writer
-	limit  uint
-	offset uint
+	limit  uint64
+	offset uint64
 }
 
-// NewTruncatedWriter returns a new truncated writer.  It wraps the given writer
-// w the way it writes up to the given limit and then ignores the rest.
-func NewTruncatedWriter(w io.Writer, limit uint) (tw *TruncatedWriter) {
+// NewTruncatedWriter returns a new truncated writer.  It wraps w so that it
+// writes up to limit bytes and then ignores the rest.
+func NewTruncatedWriter(w io.Writer, limit uint64) (tw *TruncatedWriter) {
 	return &TruncatedWriter{
 		w:     w,
 		limit: limit,
@@ -31,7 +31,7 @@ func (w *TruncatedWriter) Write(b []byte) (n int, err error) {
 		return n, nil
 	}
 
-	idx := min(uint(n), remaining)
+	idx := min(uint64(n), remaining)
 
 	// TODO(e.burkov): As the actual number of written bytes could be less then
 	// idx, consider returning this actual number.
